// headapps/intralox-sites/src/utils/enum.ts
export enum MediaType {
  IMAGE = "Image",
  VIDEO = "Video",
}

export enum ArticleType {
  NEWS = "News",
  INSIGHT = "Insights",
  SPOTLIGHT = "Spotlight",
  DOCUMENT = "Document",
}

export enum SearchCardType {
  CARD = "card",
  STANDALONE = "standalone",
  BELT_SERIES_PAGE = "beltseriespage",
  LIST_VIEW = "list",
}

export enum BeltLandingLinkLabel {
  BELTS = "Belts",
  SPROCKETS = "Sprockets",
  ACCESSORIES = "Accessories & Components",
  TOOLS = "Tools",
}

export enum TemplateName {
  SPROCKETS = "Sprockets",
  ACCESSORIES = "Accessories",
  TOOLS = "Tools",
  BELTS = "Belt",
}

export enum BeltComponentType {
  SPROCKETS = "Sprocket",
}

export enum ComponentType {
  PRODUCTS = "Products",
  INDUSTRIES = "Industries",
  SOLUTIONS = "Solutions",
}

export enum Unit {
  INCH = "us",
  MILLIMETER = "metric",
}

export enum UnitName {
  US_UNITS = "U.S. Units",
  METRIC_UNITS = "Metric Units",
}

export enum LangCountry {
  EN_US = "en_us",
  DE_DE = "de_de",
  ES_ES = "es_es",
  FR_FR = "fr_fr",
  IT_IT = "it_it",
  JA_JP = "ja_jp",
  PL_PL = "pl_pl",
  PT_BR = "pt_br",
  ZH_CN = "zh_cn",
}

export enum PaginationButton {
  NEWER = "Newer",
  OLDER = "Older",
  PREVIOUS = "Previous",
  NEXT = "Next",
}

export enum BackgroundColor {
  WHITE = "white",
  GRAY = "gray",
}